jsx refactoring to avoid code duplication
